refactor: Star, StarRating 컴포넌트 props size 추가 및 Star 컴포넌트 네이밍 수정 (#539)
