+++ frontend/src/components/@common/Star/HalfStar.tsx
@@ -4,26 +4,28 @@ import EmptyLeftStarIcon from '~/assets/icons/star/empty-left-star-icon.svg';
 import EmptyRightStarIcon from '~/assets/icons/star/empty-right-star-icon.svg';
 import FilledLeftStarIcon from '~/assets/icons/star/filled-left-star-icon.svg';
 import FilledRightStarIcon from '~/assets/icons/star/filled-right-star-icon.svg';
-import { StarRate } from '~/components/@common/StarRating/StarRating';
+
+import type { StarRate } from '~/@types/api.types';
 
 interface HalfStarProps {
   isLeft: boolean;
   isFilled: boolean;
   rateNumber: StarRate;
+  size: `${number}px`;
   onRateClick: React.MouseEventHandler<HTMLButtonElement>;
 }
 
-function HalfStar({ isLeft, isFilled, rateNumber, onRateClick }: HalfStarProps) {
+function HalfStar({ isLeft, isFilled, rateNumber, size, onRateClick }: HalfStarProps) {
   if (isLeft) {
     return (
-      <StyleHalfStarButton type="button" data-rate={rateNumber} onClick={onRateClick}>
+      <StyleHalfStarButton type="button" data-rate={rateNumber} size={size} onClick={onRateClick}>
         {isFilled ? <FilledLeftStarIcon /> : <EmptyLeftStarIcon />}
       </StyleHalfStarButton>
     );
   }
 
   return (
-    <StyleHalfStarButton type="button" onClick={onRateClick} data-rate={rateNumber}>
+    <StyleHalfStarButton type="button" onClick={onRateClick} size={size} data-rate={rateNumber}>
       {isFilled ? <FilledRightStarIcon /> : <EmptyRightStarIcon />}
     </StyleHalfStarButton>
   );
@@ -31,11 +33,15 @@ function HalfStar({ isLeft, isFilled, rateNumber, onRateClick }: HalfStarProps) 
 
 export default HalfStar;
 
-const StyleHalfStarButton = styled.button`
+const StyleHalfStarButton = styled.button<{ size: `${number}px` }>`
   padding: 0;
   margin: 0;
 
   border: none;
   background: none;
   outline: none;
+
+  & > svg {
+    width: ${({ size }) => size};
+  }
 `;
